+++ src/main/js/pathUtil.js
@@ -1,0 +1,8 @@
+//@flow
+export function pathWithTrailingSlash(path: string) {
+    const lastChar = path.substr(-1);
+    if (lastChar !== '/') {
+        return path + '/';
+    }
+    return path;
+}
