--- src/main/js/util.js
@@ -1,8 +1,0 @@
-
-export function pathWithTrailingSlash(path) {
-    var lastChar = path.substr(-1);
-    if (lastChar !== '/') {
-        path = path + '/';
-    }
-    return path;
-}
